Convert sample app to Compose Multiplatform (KMP)
- Rewrite sample/build.gradle.kts with kotlin.multiplatform, iOS targets
  (x64, arm64, simulatorArm64) exporting SampleApp framework
- Extract Example enum, navigation logic, and theme into commonMain
- Add expect/actual PlatformBackHandler (BackHandler on Android, no-op on iOS)
- Slim MainActivity to thin shell: enableEdgeToEdge + setContent { App() }
- Add MainViewController.kt as iOS entry point via ComposeUIViewController
- Move all 12 example screens unchanged to commonMain (pure Compose)
- Move Android resources (manifest, res/) to androidMain
- Add iosApp/ Xcode project with pre-build Gradle script
- Simplify Theme.kt: remove Android window APIs, keep MaterialTheme only
