+++ gradle.properties
@@ -14,6 +14,7 @@ android.nonTransitiveRClass=true
 
 # Kotlin Multiplatform
 kotlin.mpp.androidSourceSetLayoutVersion=2
+kotlin.apple.xcodeCompatibility.nowarn=true
 
 # Library version
 VERSION_NAME=1.0.0
