+++ iosApp/iosApp/iOSApp.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@main
+struct iOSApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
