+++ sample/src/androidMain/AndroidManifest.xml
