+++ sample/src/androidMain/kotlin/com/aditlal/sample/MainActivity.kt
@@ -1,0 +1,16 @@
+package com.aditlal.sample
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            App()
+        }
+    }
+}
