+++ sample/src/androidMain/kotlin/com/aditlal/sample/PlatformBackHandler.android.kt
@@ -1,0 +1,9 @@
+package com.aditlal.sample
+
+import androidx.activity.compose.BackHandler
+import androidx.compose.runtime.Composable
+
+@Composable
+internal actual fun PlatformBackHandler(enabled: Boolean, onBack: () -> Unit) {
+    BackHandler(enabled = enabled) { onBack() }
+}
