+++ sample/src/androidMain/res/drawable/ic_launcher_foreground.xml
