+++ sample/src/androidMain/res/mipmap-anydpi-v26/ic_launcher_round.xml
