+++ sample/src/androidMain/res/values/colors.xml
