+++ sample/src/commonMain/kotlin/com/aditlal/sample/PlatformBackHandler.kt
@@ -1,0 +1,6 @@
+package com.aditlal.sample
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal expect fun PlatformBackHandler(enabled: Boolean, onBack: () -> Unit)
