+++ sample/src/iosMain/kotlin/com/aditlal/sample/MainViewController.kt
@@ -1,0 +1,5 @@
+package com.aditlal.sample
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController { App() }
