+++ sample/src/iosMain/kotlin/com/aditlal/sample/PlatformBackHandler.ios.kt
@@ -1,0 +1,9 @@
+package com.aditlal.sample
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal actual fun PlatformBackHandler(enabled: Boolean, onBack: () -> Unit) {
+    // No-op: iOS has no system back button.
+    // Each example has a TopAppBar back arrow for navigation.
+}
